CreatePersonCommandTest, DeletePersonCommandTest,UpdatePersonCommandTest classes added for testing command query person table in contactdb.

+++ ContactService.Application/Persons/Commands/DeletePersonCommand.cs
@@ -1,5 +1,6 @@
 ﻿using ContactService.Application.Common.Interfaces;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 
 namespace ContactService.Application.Persons.Commands
 {
@@ -15,11 +16,13 @@ namespace ContactService.Application.Persons.Commands
 
         public async Task<Unit> Handle(DeletePersonCommand request, CancellationToken cancellationToken)
         {
-            var entity = await _context.Persons
-           .FindAsync(new object[] { request.Id }, cancellationToken);
+            if(request==null)
+                throw new ArgumentException("There's nothing to delete.");
+
+            var entity = await _context.Persons.FirstOrDefaultAsync(m => m.Id == request.Id);
 
             if (entity == null)
-                throw new ArgumentException("There's nothing to delete");
+                throw new ArgumentException("There's nothing to delete.");
 
 
             _context.Persons.Remove(entity);
